% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_units.R
\name{accumulate_contexts}
\alias{accumulate_contexts}
\title{accumulate_contexts}
\usage{
accumulate_contexts(
  x,
  codes,
  decay.function = decay(simple_window, window_size = 4),
  time.column = NULL,
  mode.column = NULL,
  mask = NULL,
  weight.by = sqrt,
  norm.by = `_sphere_norm`,
  meta.data = NULL,
  return.dena.set = FALSE,
  return.ena.set = TRUE,
  context_filter = NULL,
  summarize_ground_using = colSums,
  calculate_adj_vectors_using = ground_response_crossprod,
  ground_effect_function = function(x, y) crossprod(t(x), y),
  accumulate_unit_vectors_by = colSums
)
}
\arguments{
\item{x}{TBD}

\item{codes}{TBD}

\item{decay.function}{TBD}

\item{time.column}{TBD}

\item{mode.column}{TBD}

\item{mask}{TBD}

\item{weight.by}{TBD}

\item{norm.by}{TBD}

\item{meta.data}{TBD}

\item{return.dena.set}{TBD}

\item{return.ena.set}{TBD}

\item{context_filter}{TBD}

\item{summarize_ground_using}{TBD}

\item{calculate_adj_vectors_using}{TBD}

\item{ground_effect_function}{TBD}

\item{accumulate_unit_vectors_by}{TBD}
}
\value{
ENA set
}
\description{
accumulate_contexts
}
