% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accumulate_networks}
\alias{accumulate_networks}
\title{fast accumulate networks}
\usage{
accumulate_networks(
  x,
  code_cols,
  decay_function,
  time_col = -1L,
  ordered = TRUE
)
}
\arguments{
\item{x}{TBD}

\item{code_cols}{TBD}

\item{decay_function}{TBD}

\item{time_col}{TBD}

\item{ordered}{TBD}
}
\description{
fast accumulate networks
}
