% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threaded_ena.R
\name{accumulate_threads}
\alias{accumulate_threads}
\title{accumulate_threads}
\usage{
accumulate_threads(
  data,
  units_by,
  conversation_rules,
  code_cols,
  ...,
  conversation_splits = NULL,
  as_directed = FALSE,
  window_size = 4,
  meta_data = units_by
)
}
\arguments{
\item{data}{TBD}

\item{units_by}{TBD}

\item{conversation_rules}{TBD}

\item{code_cols}{TBD}

\item{...}{TBD}

\item{conversation_splits}{TBD}

\item{as_directed}{TBD}

\item{window_size}{TBD}

\item{meta_data}{TBD}
}
\value{
TBD
}
\description{
accumulate_threads
}
