% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjacency_key}
\alias{adjacency_key}
\title{Adjacency Key}
\usage{
adjacency_key(codes, upper = TRUE)
}
\arguments{
\item{codes}{character vector}

\item{upper}{logical, default, TRUE, returns the key for the upper triangle, FALSE - not implemented yet}
}
\value{
matrix
}
\description{
Adjacency Key
}
