% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.matrix.network.connections}
\alias{as.matrix.network.connections}
\title{Convert Network Connections to Matrix (S3 method)}
\usage{
\method{as.matrix}{network.connections}(x, ...)
}
\arguments{
\item{x}{An object of class "network.connections" (or compatible data.table/data.frame) containing connection columns (of class "network.connection").}

\item{...}{Additional arguments passed to `as.matrix`.}
}
\value{
A numeric matrix of network connections (rows = units/contexts, columns = connections).
}
\description{
This S3 method extracts the connection columns from a network connections object and returns them as a numeric matrix. It is used to facilitate matrix operations on network connection data.
}
