% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.network.connection}
\alias{as.network.connection}
\title{Re-class vector as network.connection}
\usage{
as.network.connection(x)
}
\arguments{
\item{x}{Vector to re-class}
}
\value{
re-classed vector
}
\description{
Re-class vector as network.connection
}
