% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qedata_classes.R
\name{as.qe.metadata}
\alias{as.qe.metadata}
\title{Convert a vector to 'qe.metadata' class}
\usage{
as.qe.metadata(x)
}
\arguments{
\item{x}{A vector. The vector to be converted to 'qe.metadata' class.}
}
\value{
The modified vector with the 'qe.metadata' class.
}
\description{
This function converts a vector to the 'qe.metadata' class. If the vector is a factor, it is first converted to a character vector.
}
\examples{
vec <- factor(c("A", "B", "C"))
vec <- as.qe.metadata(vec)
class(vec) # Should show 'qe.metadata' along with other classes
}
