% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.undirected.vector}
\alias{as.undirected.vector}
\title{Extract Upper Triangular Elements}
\usage{
as.undirected.vector(x, diag = FALSE)
}
\arguments{
\item{x}{A numeric matrix from which to extract upper triangular elements.}

\item{diag}{A logical value indicating whether to include the diagonal
elements. Defaults to FALSE.}
}
\value{
A vector containing the upper triangular elements of the matrix.
}
\description{
This function extracts the elements from the upper triangular part of a
matrix.
}
