% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.unordered.ordered.ena.connections}
\alias{as.unordered.ordered.ena.connections}
\title{Unorder Connections in a Matrix}
\usage{
\method{as.unordered}{ordered.ena.connections}(x)
}
\arguments{
\item{x}{A matrix or data frame containing the connections. The input should be a square matrix.}
}
\value{
A data.table with ordered connections, reclassified as "unordered.ena.connections", "ena.connections", and "ena.matrix".
}
\description{
This function takes a matrix and creates an unordered version of its connections, combining upper and lower triangular elements.
}
