% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.unordered.ordered.row.connections}
\alias{as.unordered.ordered.row.connections}
\title{Convert Ordered Row Connections to Unordered (S3 method)}
\usage{
\method{as.unordered}{ordered.row.connections}(x)
}
\arguments{
\item{x}{An object of class "ordered.row.connections" (or compatible matrix/data.frame) containing ordered connection data. The input should be a square matrix or have square number of columns.}
}
\value{
A data.table with unordered row connections, reclassified as "unordered.row.connections", "row.connections", and "ena.matrix".
}
\description{
This S3 method takes a matrix or data frame of ordered row connections (e.g., from ONA) and produces an unordered version by summing upper and lower triangular elements for each connection.
}
