% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.unordered}
\alias{as.unordered}
\title{Convert to Unordered Factor}
\usage{
as.unordered(x)
}
\arguments{
\item{x}{An object to be converted to an unordered factor.}
}
\value{
An unordered factor representation of the input object.
}
\description{
This function is a generic method to convert an object to an unordered factor.
It dispatches methods based on the class of the input object.
}
