% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{$.network.matrix}
\alias{$.network.matrix}
\title{Extract Metadata or Columns from Network Matrix (S3 method)}
\usage{
\method{$}{network.matrix}(x, i)
}
\arguments{
\item{x}{An object of class "network.matrix".}

\item{i}{Name of the column or metadata field to extract.}
}
\value{
The requested column or metadata field from the network matrix.
}
\description{
This S3 method allows convenient extraction of metadata columns from a network matrix object using the $ operator. If the requested column is metadata, it is returned from the model's meta.data; otherwise, the standard extraction is performed.
}
