% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoo.R
\name{conversation_rules}
\alias{conversation_rules}
\title{Conversation rules}
\usage{
conversation_rules(...)
}
\arguments{
\item{...}{list of rules}
}
\value{
callable expressions, see `rlang::exprs`
}
\description{
Conversation rules
}
