% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qedata_classes.R
\name{is.qe.code}
\alias{is.qe.code}
\title{Check if an object is of class 'qe.code'}
\usage{
is.qe.code(x)
}
\arguments{
\item{x}{An object. The object to be checked.}
}
\value{
A logical value. TRUE if the object is of class 'qe.code', otherwise FALSE.
}
\description{
This function checks if an object is of class 'qe.code'.
}
\examples{
dt <- 1:5
class(dt) <- c("qe.code", class(dt))
is.qe.code(dt) # Should return TRUE
}
