% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qedata_classes.R
\name{is.qe.data}
\alias{is.qe.data}
\title{Check if an object is of class 'qe.data'}
\usage{
is.qe.data(x)
}
\arguments{
\item{x}{An object. The object to be checked.}
}
\value{
A logical value. TRUE if the object is of class 'qe.data', otherwise FALSE.
}
\description{
This function checks if an object is of class 'qe.data'.
}
\examples{
library(data.table)

dt <- data.table(ID = 1:5)
class(dt) <- c("qe.data", class(dt))
is.qe.data(dt) # Should return TRUE
}
