% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qedata_classes.R
\name{is.qe.horizon}
\alias{is.qe.horizon}
\title{Check if an object is of class 'qe.horizon'}
\usage{
is.qe.horizon(x)
}
\arguments{
\item{x}{An object. The object to be checked.}
}
\value{
A logical value. TRUE if the object is of class 'qe.horizon', otherwise FALSE.
}
\description{
This function checks if an object is of class 'qe.horizon'.
}
\examples{
dt <- 1:5
class(dt) <- c("qe.horizon", class(dt))
is.qe.horizon(dt) # Should return TRUE
}
