% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qedata_classes.R
\name{is.qe.metadata}
\alias{is.qe.metadata}
\title{Check if an object is of class 'qe.metadata'}
\usage{
is.qe.metadata(x)
}
\arguments{
\item{x}{An object. The object to be checked.}
}
\value{
A logical value. TRUE if the object is of class 'qe.metadata', otherwise FALSE.
}
\description{
This function checks if an object is of class 'qe.metadata'.
}
\examples{
dt <- 1:5
class(dt) <- c("qe.metadata", class(dt))
is.qe.metadata(dt) # Should return TRUE
}
