% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qedata_classes.R
\name{is.qe.unit}
\alias{is.qe.unit}
\title{Check if an object is of class 'qe.unit'}
\usage{
is.qe.unit(x)
}
\arguments{
\item{x}{An object. The object to be checked.}
}
\value{
A logical value. TRUE if the object is of class 'qe.unit', otherwise FALSE.
}
\description{
This function checks if an object is of class 'qe.unit'.
}
\examples{
dt <- 1:5
class(dt) <- c("qe.unit", class(dt))
is.qe.unit(dt) # Should return TRUE
}
