% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_meta_data}
\alias{remove_meta_data}
\title{Remove meta columns from a data.table or data.frame}
\usage{
remove_meta_data(x)
}
\arguments{
\item{x}{A `data.table` or `data.frame` object from which meta columns should be removed.}
}
\value{
A `data.frame` with columns of class `ena.meta.data` removed.
}
\description{
This function removes columns of class `ena.meta.data` from the input object.
}
