% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{RS.data.multimodal}
\alias{RS.data.multimodal}
\title{Coded Rescushell multi-modal Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 6641 rows and 19 columns.
}
\usage{
RS.data.multimodal
}
\description{
A dataset containing sample chat data from the Rescushell Virtual
Internship with multiple modalities
}
\keyword{datasets}
