% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{RS.data}
\alias{RS.data}
\title{Coded Rescushell Chat Data}
\format{
An object of class \code{data.frame} with 3824 rows and 21 columns.
}
\usage{
RS.data
}
\description{
A dataset containing sample chat data from the Rescushell Virtual
Internship
}
\keyword{datasets}
