% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoo.R
\name{rules}
\alias{rules}
\title{Capture Subsetting Rules as Expressions}
\usage{
rules(...)
}
\arguments{
\item{...}{Logical expressions specifying the conditions for subsetting data. These expressions are captured unevaluated and returned as a list.}
}
\value{
A list of unevaluated expressions representing subsetting rules.
}
\description{
Allows users to supply conditions for subsetting rows from their data. The collected unevaluated expressions are intended to be used as the `hoo_rules` parameter in the `contexts()` function within the TMA workflow.
}
\examples{

rules(
  modality \%in\% "chat" & chatGroup \%in\% UNIT$chatGroup & condition \%in\% UNIT$condition, 
  modality \%in\% "resource" & userID \%in\% UNIT$userID & condition \%in\% UNIT$condition
)

}
