% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_mockdata}
\alias{test_mockdata}
\title{Sample Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 24 rows and 12 columns.
}
\usage{
test_mockdata
}
\description{
A small sample dataset
}
\keyword{datasets}
