% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_reddit}
\alias{test_reddit}
\title{Sample Data}
\format{
An object of class \code{data.frame} with 463 rows and 19 columns.
}
\usage{
test_reddit
}
\description{
A dataset containing threads
}
\keyword{datasets}
