% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_reddit2}
\alias{test_reddit2}
\title{Sample Data}
\format{
An object of class \code{data.frame} with 776 rows and 15 columns.
}
\usage{
test_reddit2
}
\description{
A dataset containing threads
}
\keyword{datasets}
