% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_smalldata}
\alias{test_smalldata}
\title{Sample Data}
\format{
An object of class \code{data.frame} with 12 rows and 8 columns.
}
\usage{
test_smalldata
}
\description{
A small sample dataset
}
\keyword{datasets}
