% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmodal_analysis.R
\name{tma}
\alias{tma}
\title{TMA for ENA}
\description{
TMA for ENA. rENA is used to create and visualize network models of discourse and other phenomena from coded data using Epistemic Network Analysis (ENA).
}
